# feat(serve): Opt-in parallelism for `deno serve`