--- a/cli/tools/mod.rs
+++ b/cli/tools/mod.rs
@@ -15,6 +15,7 @@ pub mod lint;
 pub mod registry;
 pub mod repl;
 pub mod run;
+pub mod serve;
 pub mod task;
 pub mod test;
 pub mod upgrade;