--- a/cli/tools/serve.rs
+++ b/cli/tools/serve.rs
@@ -1,6 +1,8 @@
 use std::sync::Arc;
 
 use deno_core::error::AnyError;
+use deno_core::futures::TryFutureExt;
+use deno_core::ModuleSpecifier;
 use deno_runtime::deno_permissions::Permissions;
 use deno_runtime::deno_permissions::PermissionsContainer;
 
@@ -9,6 +11,8 @@ use super::run::maybe_npm_install;
 use crate::args::Flags;
 use crate::args::WatchFlagsWithPaths;
 use crate::factory::CliFactory;
+use crate::util::file_watcher::WatcherRestartMode;
+use crate::worker::CliMainWorkerFactory;
 
 pub async fn serve(
   flags: Arc<Flags>,
@@ -16,8 +20,8 @@ pub async fn serve(
 ) -> Result<i32, AnyError> {
   check_permission_before_script(&flags);
 
-  if let Some(_watch_flags) = watch {
-    todo!();
+  if let Some(watch_flags) = watch {
+    return serve_with_watch(flags, watch_flags).await;
   }
 
   let factory = CliFactory::from_flags(flags);
@@ -41,6 +45,16 @@ pub async fn serve(
     &cli_options.permissions_options()?,
   )?);
   let worker_factory = factory.create_cli_main_worker_factory().await?;
+
+  do_serve(worker_factory, main_module, permissions, false).await
+}
+
+async fn do_serve(
+  worker_factory: CliMainWorkerFactory,
+  main_module: ModuleSpecifier,
+  permissions: PermissionsContainer,
+  hmr: bool,
+) -> Result<i32, AnyError> {
   let mut worker = worker_factory
     .create_main_worker(
       deno_runtime::WorkerExecutionMode::Serve,
@@ -51,25 +65,100 @@ pub async fn serve(
 
   let main = deno_core::unsync::spawn(async move { worker.run().await });
   let n = std::thread::available_parallelism()?.get();
-  for _ in 0..n {
+  let mut channels = Vec::with_capacity(n);
+  for i in 0..n {
     let worker_factory = worker_factory.clone();
     let main_module = main_module.clone();
     let permissions = permissions.clone();
+    let (tx, rx) = tokio::sync::oneshot::channel();
+    channels.push(rx);
 
-    std::thread::spawn(move || {
-      deno_runtime::tokio_util::create_and_run_current_thread(async move {
-        let mut worker = worker_factory
-          .create_main_worker(
-            deno_runtime::WorkerExecutionMode::Serve,
-            main_module,
-            permissions,
-          )
-          .await?;
-        worker.run().await?;
-        Ok::<_, AnyError>(())
-      })
-      .unwrap();
-    });
+    std::thread::Builder::new()
+      .name(format!("serve-worker-{i}"))
+      .spawn(move || {
+        deno_runtime::tokio_util::create_and_run_current_thread(async move {
+          let result =
+            run_worker(worker_factory, main_module, permissions, hmr).await;
+          let _ = tx.send(result);
+        });
+      })?;
   }
-  Ok(main.await??)
+
+  let (main_result, worker_results) = tokio::try_join!(
+    main.map_err(AnyError::from),
+    deno_core::futures::future::try_join_all(
+      channels.into_iter().map(|r| r.map_err(AnyError::from))
+    )
+  )?;
+
+  let main_ret = main_result?;
+  for res in worker_results {
+    let ret = res?;
+    if ret != 0 {
+      return Ok(ret);
+    }
+  }
+  Ok(main_ret)
+}
+
+async fn run_worker(
+  worker_factory: CliMainWorkerFactory,
+  main_module: ModuleSpecifier,
+  permissions: PermissionsContainer,
+  hmr: bool,
+) -> Result<i32, AnyError> {
+  let mut worker = worker_factory
+    .create_main_worker(
+      deno_runtime::WorkerExecutionMode::Serve,
+      main_module,
+      permissions,
+    )
+    .await?;
+  if hmr {
+    worker.run_for_watcher().await?;
+    Ok(0)
+  } else {
+    worker.run().await
+  }
+}
+
+async fn serve_with_watch(
+  flags: Arc<Flags>,
+  watch_flags: WatchFlagsWithPaths,
+) -> Result<i32, AnyError> {
+  let hmr = watch_flags.hmr;
+  crate::util::file_watcher::watch_recv(
+    flags,
+    crate::util::file_watcher::PrintConfig::new_with_banner(
+      if watch_flags.hmr { "HMR" } else { "Watcher" },
+      "Process",
+      !watch_flags.no_clear_screen,
+    ),
+    WatcherRestartMode::Automatic,
+    move |flags, watcher_communicator, _changed_paths| {
+      Ok(async move {
+        let factory = CliFactory::from_flags_for_watcher(
+          flags,
+          watcher_communicator.clone(),
+        );
+        let cli_options = factory.cli_options()?;
+        let main_module = cli_options.resolve_main_module()?;
+
+        maybe_npm_install(&factory).await?;
+
+        let _ = watcher_communicator.watch_paths(cli_options.watch_paths());
+
+        let permissions = PermissionsContainer::new(Permissions::from_options(
+          &cli_options.permissions_options()?,
+        )?);
+        let worker_factory = factory.create_cli_main_worker_factory().await?;
+
+        do_serve(worker_factory, main_module, permissions, hmr).await?;
+
+        Ok(())
+      })
+    },
+  )
+  .await?;
+  Ok(0)
 }