--- a/cli/tools/serve.rs
+++ b/cli/tools/serve.rs
@@ -1,3 +1,5 @@
+// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
+
 use std::sync::Arc;
 
 use deno_core::error::AnyError;