--- a/ext/http/00_serve.ts
+++ b/ext/http/00_serve.ts
@@ -864,9 +864,7 @@ function registerDeclarativeServer(exports) {
             );
           }
         },
-        handler: (req) => {
-          return exports.fetch(req);
-        },
+        handler: exports.fetch,
       });
     };
   }