--- a/tests/testdata/serve/parallel.ts
+++ b/tests/testdata/serve/parallel.ts
@@ -1,0 +1,7 @@
+console.log("starting serve");
+export default {
+  fetch(_req: Request) {
+    console.log("serving request");
+    return new Response("deno serve parallel");
+  },
+};